Starting on Y551 COD & Y560 NH4; add manuals
Updated URLs for all sensor models. Added placeholders as comments for some future models.

Note that Y551 replaces Y550, so I changed the model number thughout. This will not be backwards compatible, but I'm not sure anyone is using this. I might want to reconsider backward compatibility just in case.